Move Star Wars API view under separate API view that is collection of APIs.
