+++ src/app/pages/api/api.routes.config.ts
@@ -1,0 +1,12 @@
+import { RouteDetails } from '../../router';
+import { STARWARSROUTES } from './starwars/starwars.routes.config';
+
+
+export const APIROUTES: RouteDetails = {
+    url: 'api-pages',
+    name: 'api-views',
+    menuTitle: 'API Views',
+    children: {
+        starwars: STARWARSROUTES
+    }
+};
