+++ src/app/pages/api/starwars/index.ts
