+++ src/app/pages/api/starwars/planets.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, ComponentFactoryResolver, Injector } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { RouteManager } from '../../router';
+import { RouteManager } from '../../../router';
 import { StarWarsApiService, AppPlanet } from './services';
-import { getComponentHtml, RouteComponent } from '../../widgets';
+import { getComponentHtml, RouteComponent } from '../../../widgets';
 
 
 @Component({
