+++ src/app/pages/api/starwars/services/starwars-api.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { AppObservableArray, AppObservableArrayModes } from '../../../widgets';
-import { NetworkService, ConnectionOptions } from '../../../services';
+import { AppObservableArray, AppObservableArrayModes } from '../../../../widgets';
+import { NetworkService, ConnectionOptions } from '../../../../services';
 
 
 const rootUrl = 'https://swapi.co/api/';
