+++ src/app/pages/api/starwars/species-detail.component.html
