+++ src/app/pages/index.ts
@@ -2,6 +2,5 @@ export * from './about/about.component';
 export * from './home/home.component';
 export * from './header';
 export * from './footer';
-export * from './starwars';
 export { DraalAppRoutes } from './pages.routing';
 export { DraalAppPagesModule } from './pages.module';
